requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.1
fastapi==0.115.13
graphql-core==3.2.6
h11==0.16.0
idna==3.10
iso8601==2.1.0
packaging==25.0
pydantic==2.11.7
pydantic-core==2.33.2
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.275.0
tortoise-orm==0.25.1
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.34.3
websockets==15.0.1
